// packages/bootstrap-core/src/model/ConfigPreset.ts
export interface DockerServicePreset {
    ipv4_address?: string;
    openPort?: boolean | number | string;
    host?: string;
    excludeDockerService?: boolean;
    environment?: Record<string, string>;
    compose?: any;
    dockerComposeDebugMode?: boolean;
}

export interface CurrencyDistribution {
    address: string;
    amount: number;
}

export interface MosaicPreset {
    name: string;
    repeat?: number;
    main: boolean;
    harvest: boolean;
    divisibility: number;
    duration: number;
    supply: number;
    isTransferable: boolean;
    isSupplyMutable: boolean;
    isRestrictable: boolean;
    // string[] are provided public keys for unit test.
    accounts: number | string[];
    currencyDistributions: CurrencyDistribution[];
}

export interface DatabasePreset extends DockerServicePreset {
    repeat?: number;
    name: string;
    databaseName?: string;
}

export interface NemesisPreset {
    binDirectory: string;
    mosaics: MosaicPreset[];
    transactions?: Record<string, string>;
    nemesisSignerPrivateKey: string;
    transactionsDirectory: string;
}
// ...
export interface NodePreset extends DockerServicePreset, Partial<NodeConfigPreset> {
    name: string;
    harvesting: boolean;
    api: boolean;
    voting: boolean;

    repeat?: number;
    databaseHost?: string;
    host?: string;
    roles?: string;
    friendlyName?: string;
    excludeFromNemesis: boolean;

    // Optional private keys. If not provided, bootstrap will generate random ones.
    mainPrivateKey?: string;
    mainPublicKey?: string;

    transportPrivateKey?: string;
    transportPublicKey?: string;

    remotePrivateKey?: string;
    remotePublicKey?: string;

    vrfPrivateKey?: string;
    vrfPublicKey?: string;

    agentPrivateKey?: string;
    agentPublicKey?: string;

    balances?: number[];

    //Broker specific
    brokerName?: string;
    brokerHost?: string;
    brokerIpv4_address?: string;
    brokerOpenPort?: boolean | number | string;
    brokerExcludeDockerService?: boolean;
    brokerCompose?: any;
    brokerDockerComposeDebugMode?: boolean;

    //Reward program
    rewardProgram?: RewardProgram;
    rewardProgramAgentIpv4_address?: string;
    rewardProgramAgentOpenPort?: boolean | number | string;
    rewardProgramAgentExcludeDockerService?: boolean;
    rewardProgramAgentCompose?: any;
    rewardProgramAgentHost?: string;
    rewardProgramAgentDockerComposeDebugMode?: boolean;
    agentUrl?: string; //calculated if not provided.
    restGatewayUrl?: string; // calculated if not provided;
}

export interface GatewayConfigPreset {
    throttlingBurst: number;
    connectionPoolSize: number;
    apiNodeBrokerPort: number;
    apiNodeConfigPath: string;
    apiNodeBrokerPortMonitorInterval: number;
    maxConnectionAttempts: number;
    apiNodeBrokerConnectTimeout: number;
    restLoggingFilename: string;
    apiNodeBrokerMonitorLoggingThrottle: number;
    apiNodePort: number;
    throttlingRate: number;
    maxSubscriptions: number;
    apiNodeTimeout: number;
    baseRetryDelay: number;
    restDeploymentTool: string;
    restDeploymentToolVersion?: string; // default is dynamic, current bootstrap version
    restDeploymentToolLastUpdatedDate?: string; // default is dynamic, current datetime
}

export interface GatewayPreset extends DockerServicePreset, Partial<GatewayConfigPreset> {
    // At least these properties.
    repeat?: number;
    apiNodeName: string;
    apiNodeHost: string;
    databaseHost: string;
    name: string;
}

export interface ExplorerPreset extends DockerServicePreset {
    // At least these properties.
    restNodes?: string[];
    defaultNode?: string;
    repeat?: number;
    name: string;
}

export interface WalletProfilePreset {
    name: number;
    // if not provided, A file will be copied over from the working dir.
    data?: any;
    location?: string;
}

export interface WalletPreset extends DockerServicePreset {
    // At least these properties.
    repeat?: number;
    name: string;
    profiles?: WalletProfilePreset[];
    defaultNodeUrl?: string;
    restNodes?: { friendlyName: string; url: string; roles: number }[];
}

export interface FaucetPreset extends DockerServicePreset {
    // At least these properties.
    gateway: string;
    repeat?: number;
    name: string;
    privateKey?: string;
}

export interface PeerInfo {
    publicKey: string;
    endpoint: {
        host: string;
        port: number;
    };
    metadata: {
        name: string;
        roles: string;
    };
}
export type DeepPartial<T> = {
    [P in keyof T]?: DeepPartial<T[P]>;
};

export interface CommonConfigPreset extends NodeConfigPreset, GatewayConfigPreset {
    version: number; // file version
    bootstrapVersion: number;
    preset: string;
    assembly: string;
    privateKeySecurityMode?: string;
    votingKeysDirectory: string;
    sinkAddress?: string;
    epochAdjustment: string;
    catapultAppFolder: string;
    dataDirectory: string;
    subnet?: string;
    transactionsDirectory: string;
    faucetUrl?: string;
    nemesisSeedFolder?: string; // Optional seed folder if user provides an external seed/00000 folder.

    symbolWalletImage: string;
    symbolServerImage: string;
    symbolExplorerImage: string;
    symbolAgentImage: string;
    symbolRestImage: string;
    symbolFaucetImage: string;

    dockerComposeVersion: number | string;
    dockerComposeServiceRestart: string;
    dockerComposeDebugMode: boolean;
    mongoComposeRunParam: string;
    peersP2PListLimit: number;
    peersApiListLimit: number;

    nonVotingUnfinalizedBlocksDuration: string;
    votingUnfinalizedBlocksDuration?: string;
    nemesisSignerPublicKey: string;
    nemesisGenerationHashSeed: string;
    harvestNetworkFeeSinkAddress?: string;
    mosaicRentalFeeSinkAddress?: string;
    namespaceRentalFeeSinkAddress?: string;
    networkIdentifier: string;
    networkName: string;
    networkDescription: string;
    currencyMosaicId: string;
    harvestingMosaicId: string;
    baseNamespace: string;
    // This next 2 would be removed in a later merge.
    currencyName?: string;
    harvestingName?: string;
    rewardProgramEnrollmentAddress?: string;
    rewardProgramControllerApiUrl?: string;
    rewardProgramCaFile?: string;
    networkType: NetworkType;
    votingKeyDesiredLifetime: number;
    votingKeyDesiredFutureLifetime: number; // How in the future voting key files need to be generated. By default, 1 months before expiring..
    useExperimentalNativeVotingKeyGeneration?: boolean;
    lastKnownNetworkEpoch: number;
    autoUpdateVotingKeys: boolean;
    //Nested Objects
    knownRestGateways?: string[];
    inflation?: Record<string, number>;
    knownPeers?: Record<NodeType, PeerInfo[]>;
}

export interface ConfigPreset extends CommonConfigPreset {
    // Nested objects!
    nemesis: NemesisPreset;
    databases?: DatabasePreset[];
    nodes?: NodePreset[];
    gateways?: GatewayPreset[];
    explorers?: ExplorerPreset[];
    wallets?: WalletPreset[];
    faucets?: FaucetPreset[];
    customPresetCache?: CustomPreset;
}

export interface CustomPreset extends Partial<CommonConfigPreset> {
    nemesis?: DeepPartial<NemesisPreset>;
    databases?: Partial<DatabasePreset>[];
    nodes?: Partial<NodePreset>[];
    gateways?: Partial<GatewayPreset>[];
    explorers?: Partial<ExplorerPreset>[];
    wallets?: Partial<WalletPreset>[];
    faucets?: Partial<FaucetPreset>[];
}